refactor(ui): name the root-scope module for what it holds
Review feedback: `spanFilterRootScope.ts` holds nothing but constants, and a
name that reads like a module of behavior invites behavior to be added to it.
`spanFilterRootScopeConstants.ts` says what a reader will find there.
